Move dependencies into routers package.
